refactor #step2 - 패키지 구조를 변경합니다
